+++ convex/auth.config.ts
@@ -1,7 +1,7 @@
 export default {
     providers: [
       {
-        domain: "https://dynamic-badger-92.clerk.accounts.dev/",
+        domain: "https://fancy-oyster-11.clerk.accounts.dev/",
         applicationID: "convex",
       },
     ]
